dont rename copy-card. I think its fine for the module to still be copy-card to point user to the primary module. Skill then doesn't have to be updated too
